+++ src/zepben/evolve/__init__.py
@@ -90,6 +90,13 @@ from zepben.evolve.services.network.tracing.traversals.traversal import *
 from zepben.evolve.services.network.tracing.traversals.queue import *
 from zepben.evolve.services.network.tracing.traversals.branch_recursive_tracing import *
 
+from zepben.evolve.services.network.tracing.feeder.feeder_direction import *
+from zepben.evolve.services.network.tracing.util import *
+
+from zepben.evolve.services.network.translator.network_proto2cim import *
+from zepben.evolve.services.network.translator.network_cim2proto import *
+from zepben.evolve.services.network.network_service import *
+
 from zepben.evolve.services.network.tracing.connectivity import *
 from zepben.evolve.services.network.tracing.connectivity.connectivity_result import *
 from zepben.evolve.services.network.tracing.connectivity.connectivity_trace import *
@@ -99,24 +106,20 @@ from zepben.evolve.services.network.tracing.connectivity.terminal_connectivity_i
 from zepben.evolve.services.network.tracing.connectivity.transformer_phase_paths import *
 from zepben.evolve.services.network.tracing.connectivity.xy_candidate_phase_paths import *
 from zepben.evolve.services.network.tracing.connectivity.xy_phase_step import *
+from zepben.evolve.services.network.tracing.feeder.direction_status import *
 from zepben.evolve.services.network.tracing.feeder.assign_to_feeders import *
 from zepben.evolve.services.network.tracing.feeder.associated_terminal_trace import *
 from zepben.evolve.services.network.tracing.feeder.associated_terminal_tracker import *
+from zepben.evolve.services.network.tracing.feeder.set_direction import *
+from zepben.evolve.services.network.tracing.feeder.remove_direction import *
 from zepben.evolve.services.network.tracing.phases.phase_step import *
 from zepben.evolve.services.network.tracing.phases.phase_status import *
 from zepben.evolve.services.network.tracing.phases.phase_step_tracker import *
 from zepben.evolve.services.network.tracing.phases.phase_trace import *
 from zepben.evolve.services.network.tracing.phases.set_phases import *
 from zepben.evolve.services.network.tracing.phases.remove_phases import *
-from zepben.evolve.services.network.tracing.util import *
 from zepben.evolve.services.network.tracing.find import *
 from zepben.evolve.services.network.tracing.tracing import *
-from zepben.evolve.services.network.tracing.feeder.feeder_direction import *
-from zepben.evolve.services.network.tracing.feeder.direction_status import *
-
-from zepben.evolve.services.network.translator.network_proto2cim import *
-from zepben.evolve.services.network.translator.network_cim2proto import *
-from zepben.evolve.services.network.network_service import *
 
 from zepben.evolve.services.common.meta.data_source import *
 from zepben.evolve.services.common.meta.metadata_collection import *