+++ changelog.md
@@ -4,6 +4,7 @@
 ##### New Features
 * Allow specification of timeout for CimConsumerClients
 * Added convenience classes `SetPhases` and `RemovePhases` to set and remove phases on a `NetworkService`.
+* Added convenience classes `SetDirection` and `RemoveDirection` to set and remove feeder directions on a `NetworkService`.
 
 ##### Enhancements
 * BusBranchNetworkCreator logic updated so that the factory methods for topological_branches, equivalent_branches, and power_transformers get the topological nodes passed in as arguments sorted by feeder_direction.